+++ src/components/ids-data-grid/demos/in-modal.ts
@@ -1,0 +1,184 @@
+import '../../ids-modal/ids-modal';
+import '../../ids-button/ids-button';
+import type IdsDataGrid from '../ids-data-grid';
+import '../ids-data-grid';
+import type { IdsDataGridColumn } from '../ids-data-grid-column';
+import booksJSON from '../../../assets/data/books.json';
+
+const dataGrid1 = document.querySelector<IdsDataGrid>('#data-grid-1')!;
+
+(async function init() {
+  // Do an ajax request
+  const url: any = booksJSON;
+  const columns: IdsDataGridColumn[] = [];
+
+  // Set up columns
+  columns.push({
+    id: 'selectionCheckbox',
+    name: 'selection',
+    sortable: false,
+    resizable: false,
+    formatter: dataGrid1.formatters.selectionCheckbox,
+    align: 'center'
+  });
+  columns.push({
+    id: 'rowNumber',
+    name: '#',
+    formatter: dataGrid1.formatters.rowNumber,
+    sortable: false,
+    resizable: true,
+    reorderable: true,
+    readonly: true,
+    width: 65
+  });
+  columns.push({
+    id: 'description',
+    name: 'Description',
+    field: 'description',
+    sortable: true,
+    resizable: true,
+    reorderable: true,
+    formatter: dataGrid1.formatters.text,
+    filterType: dataGrid1.filters.text
+  });
+  columns.push({
+    id: 'ledger',
+    name: 'Ledger',
+    field: 'ledger',
+    resizable: true,
+    reorderable: true,
+    formatter: dataGrid1.formatters.text,
+    filterType: dataGrid1.filters.text
+  });
+  columns.push({
+    id: 'publishDate',
+    name: 'Pub. Date',
+    field: 'publishDate',
+    resizable: true,
+    reorderable: true,
+    formatter: dataGrid1.formatters.date
+  });
+  columns.push({
+    id: 'publishTime',
+    name: 'Pub. Time',
+    field: 'publishDate',
+    resizable: true,
+    reorderable: true,
+    formatter: dataGrid1.formatters.time
+  });
+  columns.push({
+    id: 'price',
+    name: 'Price',
+    field: 'price',
+    resizable: true,
+    reorderable: true,
+    formatter: dataGrid1.formatters.decimal,
+    formatOptions: { locale: 'en-US' } // Data Values are in en-US
+  });
+  columns.push({
+    id: 'bookCurrency',
+    name: 'Currency',
+    field: 'bookCurrency',
+    resizable: true,
+    reorderable: true,
+    formatter: dataGrid1.formatters.text
+  });
+  columns.push({
+    id: 'transactionCurrency',
+    name: 'Transaction Currency',
+    field: 'transactionCurrency',
+    formatter: dataGrid1.formatters.text,
+  });
+  columns.push({
+    id: 'integer',
+    name: 'Price (Int)',
+    field: 'price',
+    formatter: dataGrid1.formatters.integer,
+    formatOptions: { locale: 'en-US' } // Data Values are in en-US
+  });
+  columns.push({
+    id: 'location',
+    name: 'Location',
+    field: 'location',
+    formatter: dataGrid1.formatters.hyperlink,
+    filterType: dataGrid1.filters.text,
+    href: '#'
+  });
+  columns.push({
+    id: 'postHistory',
+    name: 'Post History',
+    field: 'postHistory',
+    formatter: dataGrid1.formatters.text,
+    filterType: dataGrid1.filters.text
+  });
+  columns.push({
+    id: 'active',
+    name: 'Active',
+    field: 'active',
+    formatter: dataGrid1.formatters.text
+  });
+  columns.push({
+    id: 'convention',
+    name: 'Convention',
+    field: 'convention',
+    formatter: dataGrid1.formatters.text
+  });
+  columns.push({
+    id: 'methodSwitch',
+    name: 'Method Switch',
+    field: 'methodSwitch',
+    formatter: dataGrid1.formatters.text
+  });
+  columns.push({
+    id: 'trackDeprecationHistory',
+    name: 'Track Deprecation History',
+    field: 'trackDeprecationHistory',
+    formatter: dataGrid1.formatters.text
+  });
+  columns.push({
+    id: 'useForEmployee',
+    name: 'Use For Employee',
+    field: 'useForEmployee',
+    formatter: dataGrid1.formatters.password
+  });
+  columns.push({
+    id: 'deprecationHistory',
+    name: 'Deprecation History',
+    field: 'deprecationHistory',
+    formatter: dataGrid1.formatters.text
+  });
+
+  dataGrid1.columns = columns;
+  const setData = async () => {
+    const res = await fetch(url);
+    const data = await res.json();
+    dataGrid1.data = data;
+  };
+
+  setData();
+
+  // Setup Modal
+  const triggerId = '#modal-trigger-btn';
+  const triggerBtn: any = document.querySelector(triggerId);
+  const modal: any = document.querySelector('ids-modal');
+
+  // Links the Modal to its trigger button (sets up click/focus events)
+  modal.target = triggerBtn;
+  modal.triggerType = 'click';
+
+  // Disable the trigger button when showing the Modal.
+  modal.addEventListener('beforeshow', () => {
+    triggerBtn.disabled = true;
+    return true;
+  });
+
+  // Close the modal when its inner button is clicked.
+  modal.onButtonClick = () => {
+    modal.hide();
+  };
+
+  // After the modal is done hiding, re-enable its trigger button.
+  modal.addEventListener('hide', () => {
+    triggerBtn.disabled = false;
+  });
+}());